refactor(llm): unify duplicated tool-message sanitization (#134)
The OpenAI-compatible and Google providers each carried a byte-identical copy of the tool-call/result adjacency sanitizer. Hoist the single provider-agnostic implementation into the llm package and have both providers call it; move its tests alongside.

Anthropic keeps its own sanitizeToolResults: it applies a broader matching policy (results matched against any preceding tool_use, not only the adjacent assistant message), so the two are not interchangeable.
