+++ internal/llm/openaicompat/convert.go
@@ -24,7 +24,7 @@ func ConvertMessages(
 	systemPrompt string,
 	convertAssistant func(msg core.Message) openai.ChatCompletionMessageParamUnion,
 ) []openai.ChatCompletionMessageParamUnion {
-	msgs = SanitizeToolMessages(msgs)
+	msgs = llm.SanitizeToolMessages(msgs)
 	msgs = DropEmptyMessages(msgs)
 	out := make([]openai.ChatCompletionMessageParamUnion, 0, len(msgs)+1)
 
@@ -51,67 +51,6 @@ func ConvertMessages(
 		}
 	}
 	return out
-}
-
-// SanitizeToolMessages ensures Chat Completions tool-call adjacency:
-// an assistant message with tool_calls must be followed immediately by tool
-// result messages for those calls. Interrupted runs and restored sessions can
-// leave orphaned tool calls/results in history; strip those before sending.
-func SanitizeToolMessages(msgs []core.Message) []core.Message {
-	result := make([]core.Message, 0, len(msgs))
-
-	for i := 0; i < len(msgs); i++ {
-		msg := msgs[i]
-
-		if msg.ToolResult != nil {
-			// Tool results are only valid immediately after their assistant call.
-			continue
-		}
-
-		if msg.Role != core.RoleAssistant || len(msg.ToolCalls) == 0 {
-			result = append(result, msg)
-			continue
-		}
-
-		j := i + 1
-		var followingResults []core.Message
-		for j < len(msgs) && msgs[j].ToolResult != nil {
-			followingResults = append(followingResults, msgs[j])
-			j++
-		}
-
-		resultIDs := make(map[string]bool, len(followingResults))
-		for _, r := range followingResults {
-			resultIDs[r.ToolResult.ToolCallID] = true
-		}
-
-		filteredCalls := make([]core.ToolCall, 0, len(msg.ToolCalls))
-		callIDs := make(map[string]bool, len(msg.ToolCalls))
-		for _, tc := range msg.ToolCalls {
-			if resultIDs[tc.ID] {
-				filteredCalls = append(filteredCalls, tc)
-				callIDs[tc.ID] = true
-			}
-		}
-
-		msg.ToolCalls = filteredCalls
-		if len(msg.ToolCalls) > 0 || msg.Content != "" {
-			result = append(result, msg)
-		}
-
-		seenResults := make(map[string]bool, len(followingResults))
-		for _, r := range followingResults {
-			id := r.ToolResult.ToolCallID
-			if callIDs[id] && !seenResults[id] {
-				result = append(result, r)
-				seenResults[id] = true
-			}
-		}
-
-		i = j - 1
-	}
-
-	return result
 }
 
 // DropEmptyMessages removes text-only messages that cannot carry provider
